Fixed product and orderitem logic

+++ src/main/java/com/example/hurtownia/domain/orderitem/OrderItemService.java
@@ -69,14 +69,15 @@ public class OrderItemService {
     public boolean delete(Long id) {
         try {
             OrderItem orderItem = findById(id);
-            if (orderItem.getOrder().getState().equals("w przygotowaniu")) {
-                Product product = productService.findById(orderItem.getProduct().getId());
-                ProductUpdateRequest productUpdateRequest = new ProductUpdateRequest();
-                BeanUtils.copyProperties(productUpdateRequest, product);
-                productUpdateRequest.setSupplierId(product.getSupplier().getId());
-                productUpdateRequest.setNumber(orderItem.getAmount() + product.getNumber());
-                productService.update(productUpdateRequest);
+            Product product = productService.findById(orderItem.getProduct().getId());
+            if (orderItem.getAmount() + product.getNumber() > product.getMaxNumber()) {
+                return false;
             }
+            ProductUpdateRequest productUpdateRequest = new ProductUpdateRequest();
+            BeanUtils.copyProperties(productUpdateRequest, product);
+            productUpdateRequest.setSupplierId(product.getSupplier().getId());
+            productUpdateRequest.setNumber(orderItem.getAmount() + product.getNumber());
+            productService.update(productUpdateRequest);
             orderItemRepository.delete(findById(id));
             return true;
         } catch (Exception e) {
@@ -122,27 +123,39 @@ public class OrderItemService {
         Product product = productService.findById(orderItemUpdateRequest.getProductId());
         ProductUpdateRequest productUpdateRequest = new ProductUpdateRequest();
         BeanUtils.copyProperties(productUpdateRequest, product);
-        if (order.getState().equals("w przygotowaniu")) {
-            if (product.getId().equals(oldOrderItem.getProduct().getId())) {
-                int amount = Math.abs(oldOrderItem.getAmount() - orderItemUpdateRequest.getAmount());
-                if (oldOrderItem.getAmount() - orderItemUpdateRequest.getAmount() >= 0) {
+        // jeśli nie zmieniamy produktu
+        if (product.getId().equals(oldOrderItem.getProduct().getId())) {
+            int amount = Math.abs(oldOrderItem.getAmount() - orderItemUpdateRequest.getAmount());
+            if (oldOrderItem.getAmount() - orderItemUpdateRequest.getAmount() >= 0) {
+                if (product.getNumber() + amount > product.getMaxNumber()) {
+                    orderItemUpdateRequest.setAmount(oldOrderItem.getAmount()-(product.getMaxNumber()-product.getNumber()));
+                    productUpdateRequest.setNumber(product.getMaxNumber());
+                }
+                else {
                     productUpdateRequest.setNumber(product.getNumber() + amount);
+                }
+            } else {
+                if (product.getNumber() < amount) {
+                    orderItemUpdateRequest.setAmount(product.getNumber() + oldOrderItem.getAmount());
+                    productUpdateRequest.setNumber(0);
                 } else {
                     productUpdateRequest.setNumber(product.getNumber() - amount);
                 }
-                productUpdateRequest.setSupplierId(product.getSupplier().getId());
-                productService.update(productUpdateRequest);
-            } else {
-                Product oldProduct = oldOrderItem.getProduct();
-                BeanUtils.copyProperties(productUpdateRequest, product);
-                productUpdateRequest.setSupplierId(product.getSupplier().getId());
-                productUpdateRequest.setNumber(product.getNumber() - orderItemUpdateRequest.getAmount());
-                productService.update(productUpdateRequest);
-                BeanUtils.copyProperties(productUpdateRequest, oldProduct);
-                productUpdateRequest.setNumber(oldProduct.getNumber() + oldOrderItem.getAmount());
-                productUpdateRequest.setSupplierId(oldProduct.getSupplier().getId());
-                productService.update(productUpdateRequest);
             }
+            productUpdateRequest.setSupplierId(product.getSupplier().getId());
+            productService.update(productUpdateRequest);
+        }
+        // gdy zmieniamy produkt
+        else {
+            Product oldProduct = oldOrderItem.getProduct();
+            BeanUtils.copyProperties(productUpdateRequest, product);
+            productUpdateRequest.setSupplierId(product.getSupplier().getId());
+            productUpdateRequest.setNumber(product.getNumber() - orderItemUpdateRequest.getAmount());
+            productService.update(productUpdateRequest);
+            BeanUtils.copyProperties(productUpdateRequest, oldProduct);
+            productUpdateRequest.setNumber(oldProduct.getNumber() + oldOrderItem.getAmount());
+            productUpdateRequest.setSupplierId(oldProduct.getSupplier().getId());
+            productService.update(productUpdateRequest);
         }
         OrderItem orderItem = findById(orderItemUpdateRequest.getId());
         orderItem.setOrder(order);
